requirements.txt
pycrypto==2.6.1
scipy==0.17.0
numpy==1.11.0

future==0.15.2